% Auto-generated: do not edit by hand
\name{chunkUpload}

\alias{chunkUpload}

\title{ChunkUpload component}

\description{

}

\usage{
chunkUpload(id=NULL, file_info=NULL, progress=NULL)
}

\arguments{
\item{id}{Character. }

\item{file_info}{Named list. }

\item{progress}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

