% Auto-generated: do not edit by hand
\name{chunkUpload}

\alias{chunkUpload}

\title{ChunkUpload component}

\description{

}

\usage{
chunkUpload(id=NULL, chunk_size=NULL, className=NULL, file_info=NULL,
file_infos=NULL, multiple=NULL, progress=NULL, style=NULL)
}

\arguments{
\item{id}{Character. }

\item{chunk_size}{Numeric. }

\item{className}{Character. }

\item{file_info}{Named list. }

\item{file_infos}{Unnamed list. }

\item{multiple}{Logical. }

\item{progress}{Numeric. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

